Enriched Inference + creating Metrics class + preparing for tasks class
