+++ scvi/inference/__init__.py
@@ -1,0 +1,7 @@
+from .inference import Inference
+from .classifier_inference import ClassifierInference
+from .variational_inference import VariationalInference, AlternateSemiSupervisedVariationalInference, \
+    JointSemiSupervisedVariationalInference
+
+all = ['Inference',
+       'ClassifierInference']
