+++ scvi/metrics/clustering.py
@@ -11,16 +11,15 @@ def get_latent_mean(vae, data_loader, use_cuda=True):
 
 @no_grad()
 @eval_modules()
-def get_latent(infer, data_loader, use_cuda=True):
+def get_latent(model, data_loader, use_cuda=True):
     latent = []
     batch_indices = []
     labels = []
-    vae = infer.model
     for tensors in data_loader:
         tensors = to_cuda(tensors, use_cuda=use_cuda)
         sample_batch, local_l_mean, local_l_var, batch_index, label = tensors
         sample_batch = sample_batch.type(torch.float32)
-        latent += [vae.sample_from_posterior_z(sample_batch, y=label)]
+        latent += [model.sample_from_posterior_z(sample_batch, y=label)]
         batch_indices += [batch_index]
         labels += [label]
     return np.array(torch.cat(latent)), np.array(torch.cat(batch_indices)), np.array(torch.cat(labels)).ravel()
