+++ scvi/metrics/log_likelihood.py
@@ -8,15 +8,14 @@ from scvi.utils import to_cuda, no_grad, eval_modules
 
 @no_grad()
 @eval_modules()
-def compute_log_likelihood(infer, data_loader, use_cuda=True):
+def compute_log_likelihood(model, data_loader, use_cuda=True):
     # Iterate once over the data_loader and computes the total log_likelihood
-    vae = infer.model
     log_lkl = 0
     for i_batch, tensors in enumerate(data_loader):
         tensors = to_cuda(tensors, use_cuda=use_cuda)
         sample_batch, local_l_mean, local_l_var, batch_index, labels = tensors
         sample_batch = sample_batch.type(torch.float32)
-        reconst_loss, kl_divergence = vae(sample_batch, local_l_mean, local_l_var, batch_index=batch_index,
+        reconst_loss, kl_divergence = model(sample_batch, local_l_mean, local_l_var, batch_index=batch_index,
                                           y=labels)
         log_lkl += torch.sum(reconst_loss).item()
     n_samples = (len(data_loader.dataset)
